Added authorization for Papers and Conferences

+++ app/controllers/conferences_controller.rb
@@ -1,31 +1,41 @@
 class ConferencesController < ApplicationController
   before_action :set_conference, only: [:show, :edit, :update, :destroy]
+  before_action :authenticate_account!
 
+  def pundit_user
+    current_account
+  end
   # GET /conferences
   # GET /conferences.json
   def index
+    authorize Conference
+    @conferences = policy_scope(Conference)
+
     @conferences = Conference.all
   end
 
   # GET /conferences/1
   # GET /conferences/1.json
   def show
+    authorize @conference
   end
 
   # GET /conferences/new
   def new
     @conference = Conference.new
+    authorize @conference
   end
 
   # GET /conferences/1/edit
   def edit
+    authorize @conference
   end
 
   # POST /conferences
   # POST /conferences.json
   def create
     @conference = Conference.new(conference_params)
-
+    authorize @conference
     respond_to do |format|
       if @conference.save
         format.html { redirect_to @conference, notice: 'Conference was successfully created.' }
@@ -40,6 +50,7 @@ class ConferencesController < ApplicationController
   # PATCH/PUT /conferences/1
   # PATCH/PUT /conferences/1.json
   def update
+    authorize @conference
     respond_to do |format|
       if @conference.update(conference_params)
         format.html { redirect_to @conference, notice: 'Conference was successfully updated.' }
@@ -54,8 +65,9 @@ class ConferencesController < ApplicationController
   # DELETE /conferences/1
   # DELETE /conferences/1.json
   def destroy
+    authorize @conference
     @conference = Conference.find(params[:id])
-   @Conference.papers.each do |paper| 
+    @Conference.papers.each do |paper| 
       if(paper.conference == @Conference)
         paper.destroy
       end
